Move download tools dependency (#203)
* create base api call for address parsing

* modified pylint file to accept pydantic basemodel

* added Dockerfile app image definition

* added tests for the api for single address

* linted package

* added cache logic to docker container

* moved download tools dependencies

* moved download tools dependencies

* removed unwanted changes

* moved download tools tests to its correct test file

* moved downloads code before fasttext licence statement

* resolved comments, removed Annotated type and added docs

* Update deepparse/download_tools.py

* removed casting to list in model choices

* added unit test for wrong model type

* fixed default cache path

* fix black

* fix pylint

---------
