+++ deepparse/__init__.py
@@ -1,5 +1,5 @@
 # pylint: disable=wildcard-import
-from .fasttext_tools import *
-from .tools import *
+from .download_tools import *
+from .validations import *
 from .version import __version__
 from .weights_tools import *
