+++ deepparse/cli/download_model.py
@@ -1,17 +1,8 @@
 import argparse
-import os
 import sys
 
-from bpemb import BPEmb
 
-from deepparse import (
-    CACHE_PATH,
-    download_fasttext_magnitude_embeddings,
-    latest_version,
-    download_fasttext_embeddings,
-    download_weights,
-    MODEL_CHOICES,
-)
+from deepparse.download_tools import download_model, MODEL_MAPPING_CHOICES
 
 
 def main(args=None) -> None:
@@ -32,31 +23,10 @@ def main(args=None) -> None:
     parsed_args = get_args(args)
 
     model_type = parsed_args.model_type
-    if "-attention" in model_type:
-        root_model_type = model_type.split("-")[0]
-        model_type = root_model_type + "_attention"
 
     saving_cache_path = parsed_args.saving_cache_dir
 
-    if saving_cache_path is None:
-        saving_cache_path = CACHE_PATH
-
-    if "fasttext" in model_type and "fasttext-light" not in model_type:
-        download_fasttext_embeddings(cache_dir=saving_cache_path)
-    elif model_type == "fasttext-light":
-        download_fasttext_magnitude_embeddings(cache_dir=saving_cache_path)
-    elif "bpemb" in model_type:
-        BPEmb(
-            lang="multi", vs=100000, dim=300, cache_dir=saving_cache_path
-        )  # The class manage the download of the pretrained words embedding
-
-    model_path = os.path.join(saving_cache_path, f"{model_type}.ckpt")
-    version_path = os.path.join(saving_cache_path, f"{model_type}.version")
-    if not os.path.isfile(model_path) or not os.path.isfile(version_path):
-        download_weights(model_type, saving_dir=saving_cache_path)
-    elif not latest_version(model_type, cache_path=saving_cache_path, verbose=True):
-        print("A new version of the pretrained model is available. The newest model will be downloaded.")
-        download_weights(model_type, saving_dir=saving_cache_path)
+    download_model(model_type, saving_cache_path=saving_cache_path)
 
 
 def get_parser() -> argparse.ArgumentParser:
@@ -64,7 +34,7 @@ def get_parser() -> argparse.ArgumentParser:
     parser = argparse.ArgumentParser()
     parser.add_argument(
         "model_type",
-        choices=MODEL_CHOICES,
+        choices=MODEL_MAPPING_CHOICES,
         help="The model type to download.",
     )
     parser.add_argument(
