+++ deepparse/cli/download_models.py
@@ -1,17 +1,7 @@
 import argparse
-import os
 import sys
 
-from bpemb import BPEmb
-
-from deepparse import (
-    CACHE_PATH,
-    download_fasttext_magnitude_embeddings,
-    latest_version,
-    download_fasttext_embeddings,
-    download_weights,
-    MODEL_CHOICES,
-)
+from deepparse.download_tools import download_models
 
 
 def main(args=None) -> None:
@@ -33,23 +23,7 @@ def main(args=None) -> None:
 
     saving_cache_path = parsed_args.saving_cache_dir
 
-    if saving_cache_path is None:
-        saving_cache_path = CACHE_PATH
-
-    download_fasttext_embeddings(cache_dir=saving_cache_path)
-    download_fasttext_magnitude_embeddings(cache_dir=saving_cache_path)
-    BPEmb(
-        lang="multi", vs=100000, dim=300, cache_dir=saving_cache_path
-    )  # The class manage the download of the pretrained words embedding
-
-    for model_type in MODEL_CHOICES:
-        model_path = os.path.join(saving_cache_path, f"{model_type}.ckpt")
-        version_path = os.path.join(saving_cache_path, f"{model_type}.version")
-        if not os.path.isfile(model_path) or not os.path.isfile(version_path):
-            download_weights(model_type, saving_dir=saving_cache_path)
-        elif not latest_version(model_type, cache_path=saving_cache_path, verbose=True):
-            print("A new version of the pretrained model is available. The newest model will be downloaded.")
-            download_weights(model_type, saving_dir=saving_cache_path)
+    download_models(saving_cache_path=saving_cache_path)
 
 
 def get_parser() -> argparse.ArgumentParser:
