+++ deepparse/cli/parser_arguments_adder.py
@@ -1,13 +1,13 @@
 from argparse import ArgumentParser
 
 from .tools import wrap, bool_parse
-from .. import MODEL_CHOICES
+from .. import MODEL_MAPPING_CHOICES
 
 
 def add_base_parsing_model_arg(parser: ArgumentParser) -> None:
     parser.add_argument(
         "base_parsing_model",
-        choices=MODEL_CHOICES,
+        choices=MODEL_MAPPING_CHOICES,
         help=wrap("The base parsing module to use for retraining."),
     )
