+++ tests/parser/test_address_parser_retrain_api.py
@@ -166,7 +166,7 @@ class AddressParserRetrainTest(AddressParserPredictTestCase):
 
         optimizer_mock.assert_called_with(self.model_mock.parameters(), self.a_learning_rate)
 
-    @patch("deepparse.tools.poutyne")
+    @patch("deepparse.validations.poutyne")
     @patch("deepparse.parser.address_parser.torch.save")
     @patch("deepparse.parser.address_parser.Experiment")
     @patch("deepparse.parser.address_parser.SGD")
@@ -204,7 +204,7 @@ class AddressParserRetrainTest(AddressParserPredictTestCase):
 
         self.assertEqual(actual, expected)
 
-    @patch("deepparse.tools.poutyne")
+    @patch("deepparse.validations.poutyne")
     @patch("deepparse.parser.address_parser.torch.save")
     @patch("deepparse.parser.address_parser.Experiment")
     @patch("deepparse.parser.address_parser.SGD")
